pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "django-quanttide-idam"
version = "0.1.0-alpha.1"
description = "量潮身份与访问管理Django工具箱"
authors = ["QuantTide Inc. <opensource@quanttide.com>"]
license = "Apache-2.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "Framework :: Django",
    "License :: OSI Approved :: Apache Software License",
    "Development Status :: 2 - Pre-Alpha"
]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.0"
djangorestframework = "^3.0"
authlib = "^1.0"
phonenumbers = "^8.13.18"
django-phonenumber-field = "^7.1.0"

[tool.poetry.dev-dependencies]
